[sdk] reduce circular deps
