+++ packages/sdk/src/evm/common/any-evm-utils/caches.ts
@@ -1,0 +1,5 @@
+import type { DeploymentPreset } from "../../types/any-evm/deploy-data";
+
+export const caches = {
+  deploymentPresets: {} as Record<string, DeploymentPreset>,
+};
