+++ packages/sdk/src/evm/common/any-evm-utils/getSaltHash.ts
@@ -1,0 +1,16 @@
+import { ethers } from "ethers";
+
+/**
+ * Generate salt for deployment with Create2
+ * Note: Salt component is generated by appending `tw` (thirdweb) to the bytecode
+ *
+ * @internal
+ * @param bytecode: Creation bytecode of the contract to deploy
+ */
+export function getSaltHash(bytecode: string): string {
+  const bytecodeHash = ethers.utils.id(bytecode);
+  const salt = `tw.${bytecodeHash}`;
+  const saltHash = ethers.utils.id(salt);
+
+  return saltHash;
+}
