+++ packages/sdk/src/evm/common/any-evm-utils/getThirdwebContractAddress.ts
@@ -1,0 +1,32 @@
+import { ThirdwebStorage } from "@thirdweb-dev/storage";
+import invariant from "tiny-invariant";
+import { getChainProvider } from "../../constants/urls";
+import { DEFAULT_API_KEY } from "../../../core/constants/urls";
+import { isContractDeployed } from "./isContractDeployed";
+import { predictThirdwebContractAddress } from "./predictThirdwebContractAddress";
+
+/**
+ *
+ * @public
+ * @param contractName
+ * @param chainId
+ * @param storage
+ */
+export async function getThirdwebContractAddress(
+  contractName: string,
+  chainId: number,
+  storage: ThirdwebStorage,
+): Promise<string> {
+  const provider = getChainProvider(chainId, {
+    thirdwebApiKey: DEFAULT_API_KEY,
+  });
+  const contractAddress = await predictThirdwebContractAddress(
+    contractName,
+    chainId,
+    storage,
+  );
+  const isDeployed = await isContractDeployed(contractAddress, provider);
+  invariant(isDeployed, "Contract not deployed yet");
+
+  return contractAddress;
+}
