+++ packages/sdk/src/evm/common/deploy.ts
@@ -18,7 +18,7 @@ import { BigNumber, Signer, providers } from "ethers";
 import { z } from "zod";
 import { ThirdwebStorage } from "@thirdweb-dev/storage";
 import { SUPPORTED_CHAIN_IDS } from "../constants/chains";
-import { computeForwarderAddress } from "./any-evm-utils";
+import { computeForwarderAddress } from "./any-evm-utils/computeForwarderAddress";
 import type {
   PrebuiltContractType,
   DeploySchemaForPrebuiltContractType,
