+++ packages/sdk/src/evm/common/index.ts
@@ -29,13 +29,42 @@ export * from "./feature-detection/hasFunction";
 export * from "./version-checker";
 export * from "./currency";
 export * from "./verification";
-export * from "./any-evm-utils";
+
+// any-evm-utils
+export * from "./any-evm-utils/constants";
+export * from "./any-evm-utils/isContractDeployed";
+export * from "./any-evm-utils/isEIP155Enforced";
+export * from "./any-evm-utils/getCreate2FactoryAddress";
+export * from "./any-evm-utils/getSaltHash";
+export * from "./any-evm-utils/getInitBytecodeWithSalt";
+export * from "./any-evm-utils/computeDeploymentAddress";
+export * from "./any-evm-utils/computeEOAForwarderAddress";
+export * from "./any-evm-utils/computeForwarderAddress";
+export * from "./any-evm-utils/computeCloneFactoryAddress";
+export * from "./any-evm-utils/computeNativeTokenAddress";
+export * from "./any-evm-utils/getThirdwebContractAddress";
+export * from "./any-evm-utils/predictThirdwebContractAddress";
+export * from "./any-evm-utils/getEncodedConstructorParamsForThirdwebContract";
+export * from "./any-evm-utils/getKeylessTxn";
+export * from "./any-evm-utils/deployCreate2Factory";
+export * from "./any-evm-utils/deployContractDeterministicRaw";
+export * from "./any-evm-utils/deployContractDeterministic";
+export * from "./any-evm-utils/getDeploymentInfo";
+export * from "./any-evm-utils/deployWithThrowawayDeployer";
+export * from "./any-evm-utils/computeDeploymentInfo";
+export * from "./any-evm-utils/encodeConstructorParamsForImplementation";
+export * from "./any-evm-utils/convertParamValues";
+export * from "./any-evm-utils/getCreate2FactoryDeploymentInfo";
+export * from "./any-evm-utils/fetchAndCachePublishedContractURI";
+export * from "./any-evm-utils/fetchAndCacheDeployMetadata";
+export * from "./any-evm-utils/estimateGasForDeploy";
+export * from "./any-evm-utils/createTransactionBatches";
+
 export * from "./deploy";
 export {
   convertToReadableQuantity,
   fetchSnapshotEntryForAddress,
 } from "./claim-conditions";
 export { getCachedAbiForContract } from "./abi";
-export * from "./any-evm-utils";
 export * from "./ens";
 export { prepareGaslessRequest } from "./transactions";
