+++ packages/sdk/src/evm/zksync/zksync-deploy-utils.ts
@@ -1,9 +1,7 @@
 import * as zk from "zksync-web3";
 import { twProxyArtifactZK } from "./temp-artifact/TWProxy";
-import {
-  fetchAndCacheDeployMetadata,
-  convertParamValues,
-} from "../common/any-evm-utils";
+import { fetchAndCacheDeployMetadata } from "../common/any-evm-utils/fetchAndCacheDeployMetadata";
+import { convertParamValues } from "../common/any-evm-utils/convertParamValues";
 import { extractConstructorParamsFromAbi } from "../common/feature-detection/extractConstructorParamsFromAbi";
 import { extractFunctionParamsFromAbi } from "../common/feature-detection/extractFunctionParamsFromAbi";
 import { BytesLike, Contract, Signer, ethers } from "ethers";
